improvements to groove
v2018.07.18
A draft path array is now used to create the new bspline above the original bspline in order to create a more uniformly sized groove.  The draft path array uses a circle of radius = slot height, and a point on each circle in the array is used as a list of points from which to create the new bspline (as opposed to simply moving its placement up the z axis).
Added precision parameter.